[fix] fix comment, drop Partition.java
