--- paimon-core/src/main/java/org/apache/paimon/catalog/Partition.java
@@ -1,56 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.paimon.catalog;
-
-import org.apache.paimon.annotation.Public;
-
-import java.util.LinkedHashMap;
-import java.util.Objects;
-
-/**
- * Partition with the specified values.
- *
- * @since 0.5.0
- */
-@Public
-public class Partition {
-    private final LinkedHashMap<String, String> partitionSpec;
-
-    public Partition(LinkedHashMap<String, String> partitionSpec) {
-        this.partitionSpec = partitionSpec;
-    }
-
-    public LinkedHashMap<String, String> partitionSpec() {
-        return partitionSpec;
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(partitionSpec);
-    }
-
-    @Override
-    public String toString() {
-        return "Partition{" + partitionSpec + '}';
-    }
-
-    public static Partition of(LinkedHashMap<String, String> spec) {
-        return new Partition(spec);
-    }
-}
