+++ paimon-core/src/main/java/org/apache/paimon/table/source/snapshot/SnapshotSplitReader.java
@@ -19,8 +19,8 @@
 package org.apache.paimon.table.source.snapshot;
 
 import org.apache.paimon.Snapshot;
-import org.apache.paimon.catalog.Partition;
 import org.apache.paimon.consumer.ConsumerManager;
+import org.apache.paimon.data.BinaryRow;
 import org.apache.paimon.operation.ScanKind;
 import org.apache.paimon.predicate.Predicate;
 import org.apache.paimon.table.source.DataSplit;
@@ -50,5 +50,5 @@ public interface SnapshotSplitReader {
     List<DataSplit> overwriteSplits();
 
     /** Get partitions from a snapshot. */
-    List<Partition> partitions();
+    List<BinaryRow> partitions();
 }
